chore: move server file
