+++ tsconfig.app.json
@@ -5,6 +5,6 @@
     "outDir": "./out-tsc/app",
     "types": ["node"]
   },
-  "files": ["src/main.ts", "src/main.server.ts", "server.ts"],
+  "files": ["src/main.ts", "src/main.server.ts", "src/server.ts"],
   "include": ["src/**/*.d.ts"]
 }
